+++ geometry/run.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-g++ -std=c++17 -I./src test/test.cpp src/geometry.cpp -o geometry
+g++ -std=c++17 -I./src test/test.cpp -o geometry
 ./geometry
 
 echo All tests passed!
